Move `ProjectionStorage` under the `projection` package
