+++ server/src/main/java/org/spine3/server/storage/StorageFactory.java
@@ -27,6 +27,7 @@ import org.spine3.server.command.CommandStore;
 import org.spine3.server.entity.Entity;
 import org.spine3.server.event.EventStorage;
 import org.spine3.server.event.EventStore;
+import org.spine3.server.projection.ProjectionStorage;
 import org.spine3.server.stand.Stand;
 
 /**
