+++ server/src/main/java/org/spine3/server/storage/memory/InMemoryStorageFactory.java
@@ -25,7 +25,7 @@ import org.spine3.server.aggregate.AggregateStorage;
 import org.spine3.server.command.CommandStorage;
 import org.spine3.server.entity.Entity;
 import org.spine3.server.event.EventStorage;
-import org.spine3.server.storage.ProjectionStorage;
+import org.spine3.server.projection.ProjectionStorage;
 import org.spine3.server.storage.RecordStorage;
 import org.spine3.server.storage.StandStorage;
 import org.spine3.server.storage.StorageFactory;
