+++ server/src/test/java/org/spine3/server/projection/ProjectionStorageShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.server.storage;
+package org.spine3.server.projection;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.FieldMask;
@@ -29,6 +29,8 @@ import org.junit.Test;
 import org.spine3.protobuf.AnyPacker;
 import org.spine3.protobuf.Durations;
 import org.spine3.protobuf.Timestamps;
+import org.spine3.server.storage.AbstractStorageShould;
+import org.spine3.server.storage.EntityStorageRecord;
 import org.spine3.test.Tests;
 import org.spine3.test.projection.Project;
 import org.spine3.test.projection.ProjectId;
