+++ server/src/test/java/org/spine3/server/storage/memory/InMemoryProjectionStorageShould.java
@@ -20,8 +20,8 @@
 
 package org.spine3.server.storage.memory;
 
-import org.spine3.server.storage.ProjectionStorage;
-import org.spine3.server.storage.ProjectionStorageShould;
+import org.spine3.server.projection.ProjectionStorage;
+import org.spine3.server.projection.ProjectionStorageShould;
 
 import static org.spine3.base.Identifiers.newUuid;
